backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
web3==6.11.0
pymongo==4.6.0
python-dotenv==1.0.0
